Bug #1257: compile warnings due to pid_t not being an int
Forward port of 2.5 patch and added use of debugs() where applicable

+++ helpers/ntlm_auth/SMB/ntlm.h
@@ -53,7 +53,7 @@
 static char *__foo;
 extern char debug_enabled;
 #define debug(X...) if (debug_enabled) { \
-                    fprintf(stderr,"ntlm-auth[%d](%s:%d): ", getpid(), \
+                    fprintf(stderr,"ntlm-auth[%ld](%s:%d): ", (long)getpid(), \
                     ((__foo=strrchr(__FILE__,'/'))==NULL?__FILE__:__foo+1),\
                     __LINE__);\
                     fprintf(stderr,X); }
